% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APT.R
\name{apt}
\alias{apt}
\title{Parallel tempering using a HMC kernel with an adaptive tuning procedure}
\usage{
apt(
  x.init,
  target,
  gd.target,
  maxiter,
  maxtemp,
  ntemp,
  njumps = 50,
  sumstat_fun = identity,
  ntemp_tuning_method = "rectangular",
  search_center,
  search_scale,
  Uthreshold,
  span = 100,
  verbose = FALSE
)
}
\arguments{
\item{x.init}{starting point for the Markov chain}

\item{target}{an R function that evaluates the log target density}

\item{gd.target}{an R function that evaluates the gradient of the log target density}

\item{maxiter}{the maximum number of MCMC iterations (the length of each parallel chain)}

\item{maxtemp}{starting value for the maximum log temperature}

\item{ntemp}{starting value for the number of parallel chains minus one (number of auxiliary chains besides the chain at temperature = 1}

\item{njumps}{number of leapfrog steps for HMC at each iteration of parallel chains}

\item{sumstat_fun}{the function to apply to compute the summary statistic to be returned. Defaults to the identity function.}

\item{ntemp_tuning_method}{method for tuning the number of parallel chains. Either "rectangular" (default) or "Uthreshold".}

\item{search_center}{the reference point for assessing whether the top level chain meets a search criterion. Used if \code{ntemp_tuning_method} is "rectangular".}

\item{search_scale}{the vector of length equal to the dimension of the space specifying the search scale for each coordinate. Used if \code{ntemp_tuning_method} is "rectangular".}

\item{Uthreshold}{the required value of the potential function U to be reached before stopping the tuning of ntemp. Used if \code{ntemp_tuning_method} is "Uthreshold".}

\item{span:}{a scalar value specifying how many past MCMC iterations at the top level should be considered when checking the search criterion.}

\item{verbose:}{logical. should detailed internal outputs be printed in buffer?}
}
\value{
A list consisting of:
\itemize{
\item{jsizes: Tuned values of leapfrog step sizes for parallel chains.}
\item{tempseq: The (tuned) temperature levels.}
\item{acc_prob: The history of acceptance probabilities for the parallel chains.}
\item{sumstat: The history of parallel chains: the values are the evaluation of \code{sumstat_fun} on the chain states.}
\item{cumnjump: Cumulative number of leapfrog steps up to the current iteration (recorded for all parallel chains).}
}
}
\description{
Parallel tempering, in which each chain is constructed using HMC. Adopts the temperature level adjustment strategy proposed by Miasojedow, Moulines, and Vihola (2013), JCGS. Additionally, the number of parallel chains is adaptively tuned, such that the top level chain satisfies a specified search criterion.
}
\references{
\itemize{
\item{Park, J. (2025). Sampling from high-dimensional, multimodal distributions using automatically-tuned, tempered Hamiltonian Monte Carlo \url{https://doi.org/10.48550/arXiv.2111.06871}}
\item{Miasojedow, B., Moulines, E., and Vihola, M. (2013). An adaptive parallel tempering algorithm. Journal of Computational and Graphical Statistics, 22(3):649–664.}
}
}
