% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATSMC.R
\name{ATSMC}
\alias{ATSMC}
\title{Tempered sequential Monte Carlo (TSMC) with adaptive tuning}
\usage{
ATSMC(
  logtarget,
  gradlt,
  sumstat = identity,
  base_gen,
  base_lpdf,
  base_glpdf,
  nparticle,
  adaptune = TRUE,
  targetESSratio = 0.5,
  invtemp_schedule,
  ntemp = 15,
  lfstepsize = 0.1,
  n_lfsteps = 10,
  MCMCsteps = 5,
  bounds = NULL,
  return_all = FALSE,
  verbose = FALSE,
  parallel = FALSE,
  ncore
)
}
\arguments{
\item{logtarget}{log target density function (R function)}

\item{gradlt}{gradient of log target density (R function)}

\item{sumstat}{an R function that is applied to the vector \code{x} to create a summary statistic for each MCMC iteration for output. Here \code{x} is the current state of the Markov chain. The return value should be a vector. The default is the identity function. This argument is useful for avoiding high memory usage when \code{x} is high dimensional.}

\item{base_gen}{An R function that generates a single draw from the base distribution. This function does not take any argument.}

\item{base_lpdf}{The log probability density function of the base distribution, p_0(x). Intermediate tempered densities are given by p_0(x)^(1-beta) * p(x)^beta, where p(x) is the target density, and beta is the inverse temperature.}

\item{base_glpdf}{The gradient of the log probability density function of the base distribution, log(p_0(x)) (see the description for `base_lpdf').}

\item{nparticle}{number of particles to be used.}

\item{adaptune}{logical. Should \verb{ntemp' and }tempschedule' be tuned adaptively? If TRUE, both parameters will be tuned such that the effective sample size (ESS) for the importance sampling step for adjacent temperature levels is close to the specified \verb{targetESSratio' times }nparticle'.}

\item{targetESSratio}{The target ratio of the effective sample size to number of particles for adaptively tuning `invtemp_schedule'.}

\item{invtemp_schedule}{`invtemp_schedule' gives a numeric vector of inverse temperature (beta) levels. It is expected that the vector forms an increasing sequence, where the first entry equals 0 and the last entry equals 1. If adaptune=TRUE, this argument will be ignored.}

\item{ntemp}{The number of temperature levels, including the base (beta=0) and the target (beta=1). If adaptune=FALSE, \verb{invtemp_schedule' will be a uniform sequence between 0 and 1 of length }ntemp'. If 'invtemp_schedule' is provided, 'ntemp' will be ignored.}

\item{lfstepsize}{Starting value for tuning the leapfrog step size for the HMC kernel used in between importance sampling steps to perturb parameters (and replenish particle diversity). The leapfrog step size is tuned so that the acceptance rate is close to 0.9.}

\item{n_lfsteps}{Base number of leapfrog steps for the HMC kernel. Repeated until the correlation is close to zero (see Buchholz, Chopin, Jacob (2021) for details.)}

\item{MCMCsteps}{Number of HMC perturbations carried out in sequence for each particle at each temperature level}

\item{bounds}{the bounds for the variables (used for constructing HMC trajectories). See '?leapfrog' for more details.}

\item{return_all:}{logical. If TRUE, return (the summary statistics of) all particles at all temperature levels}

\item{verbose:}{logical. If TRUE, diagnostic messages are printed.}

\item{parallel:}{logical. Is particle rejuvenation step carried out in parallel using multiple cores? Defaults to TRUE.}

\item{ncore:}{optional. The number of computing cores to use for parallel particle rejuvenation. Automatically capped at the number of available cores minus one.}
}
\value{
A return list. If \code{return_all} is FALSE, the list consists of:
\itemize{
\item{particles: The ensemble of particles at the final temperature level. These particles provide a Monte Carlo approximation to the target distribution. Note that the return values are the values of the function \code{sumstat} applied to the particles. Thus \code{particles} do not return the actual particles unless \code{sumstat} is the identity function (default).}
\item{weights: The normalized weights for the particles at the final step. Note: The \code{weights} vector does not correspond to the \code{particles} vector, in the sense that \code{particles} give the values after importance resampling with \code{weights}. Thus \code{weights} are provided mainly for diagnostic purposes.}
\item{invtemp_schedule: The sequence of inverse temperature levels. If \code{adaptune} is TRUE, the tuned sequence of values will be returned. Otherwise, the provided sequence will be returned.}
}
If \code{return_all} is TRUE, the list additionally consists of
\itemize{
\item{particles_book: A list consisting of particle ensembles \emph{before} the importance resampling steps at all temperature levels. The first list entry is a collection of iid draw from the base distribution (\code{base_gen}).}
\item{weights_book: A list of normalized weights at all temperature levels.}
}
}
\description{
Tempered sequential Monte Carlo is also known as \emph{anealed importance sampling}. The adaptive tuning procedure mostly follow the methods proposed by Buchholz et al. (2021).
}
\references{
\itemize{
\item{Buchholz, A., Chopin, N., and Jacob, P. E. (2021). Adaptive tuning of Hamiltonian Monte Carlo within sequential Monte Carlo. Bayesian Analysis, 16(3):745–771}
\item{Neal, R. M. (2001). Annealed importance sampling. Statistics and Computing, 11:125–139.}
\item{Park, J. (2025). Sampling from high-dimensional, multimodal distributions using automatically-tuned, tempered Hamiltonian Monte Carlo \url{https://doi.org/10.48550/arXiv.2111.06871}}
}
}
