% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leapfrog.R
\name{lf}
\alias{lf}
\title{Leapfrog function}
\usage{
lf(x, v, gd, jsize, njumps, massInv = 1)
}
\arguments{
\item{x}{a vector of initial position}

\item{v}{a vector of initial velocity}

\item{gd}{an R function that returns the gradient of log target density}

\item{jsize}{leapfrog step size}

\item{njumps}{number of leapfrog jumps}

\item{massInv}{mass-inverse matrix (currently, it is allowed to be only a scalar or a vector of the same length as v. When scalar, this means that the mass-inverse matrix is that constant times the identity matrix, and if \code{massInv} is a vector, then the mass-inverse matrix is the diagonal matrix having the diagonal entries given by \code{massInv}.)}
}
\value{
A concatenated vector \code{c(x,v)} of the final position and velocity
}
\description{
carries out a certain number of leapfrog steps
}
